# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/stt/model": {
      "get": {
        "tags": [
          "speech-model"
        ],
        "summary": "Get Speech Model Status",
        "operationId": "get_speech_model_status_api_stt_model_get",
        "parameters": [
          {
            "name": "backend",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "vosk",
                "whisper"
              ],
              "type": "string",
              "title": "Backend"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "ja",
                "en"
              ],
              "type": "string",
              "title": "Language"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "tiny",
                    "base",
                    "small",
                    "medium",
                    "large-v2",
                    "large-v3-turbo"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeechModelStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stt/model/download": {
      "post": {
        "tags": [
          "speech-model"
        ],
        "summary": "Start Speech Model Download",
        "operationId": "start_speech_model_download_api_stt_model_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechModelDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeechModelStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stt/model/cancel": {
      "post": {
        "tags": [
          "speech-model"
        ],
        "summary": "Cancel Speech Model Download",
        "operationId": "cancel_speech_model_download_api_stt_model_cancel_post",
        "parameters": [
          {
            "name": "backend",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "vosk",
                "whisper"
              ],
              "type": "string",
              "title": "Backend"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "ja",
                "en"
              ],
              "type": "string",
              "title": "Language"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "tiny",
                    "base",
                    "small",
                    "medium",
                    "large-v2",
                    "large-v3-turbo"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeechModelStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpeechModelDownloadRequest": {
        "properties": {
          "backend": {
            "type": "string",
            "enum": [
              "vosk",
              "whisper"
            ],
            "title": "Backend"
          },
          "language": {
            "type": "string",
            "enum": [
              "ja",
              "en"
            ],
            "title": "Language"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "tiny",
                  "base",
                  "small",
                  "medium",
                  "large-v2",
                  "large-v3-turbo"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "backend",
          "language"
        ],
        "title": "SpeechModelDownloadRequest",
        "description": "Request a provider-controlled model without accepting arbitrary URLs or paths."
      },
      "SpeechModelStatusResponse": {
        "properties": {
          "backend": {
            "type": "string",
            "enum": [
              "vosk",
              "whisper"
            ],
            "title": "Backend"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "missing",
              "downloading",
              "ready",
              "failed",
              "cancelled"
            ],
            "title": "State"
          },
          "phase": {
            "type": "string",
            "enum": [
              "idle",
              "downloading",
              "verifying",
              "extracting",
              "ready"
            ],
            "title": "Phase"
          },
          "language": {
            "type": "string",
            "enum": [
              "ja",
              "en"
            ],
            "title": "Language"
          },
          "downloaded_bytes": {
            "type": "integer",
            "title": "Downloaded Bytes"
          },
          "total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Bytes"
          },
          "progress_percent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Percent"
          },
          "model_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Path"
          },
          "storage_path": {
            "type": "string",
            "title": "Storage Path"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "network",
                  "disk_full",
                  "permission",
                  "checksum",
                  "archive",
                  "cancelled",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "cancelable": {
            "type": "boolean",
            "title": "Cancelable"
          }
        },
        "type": "object",
        "required": [
          "backend",
          "model_id",
          "state",
          "phase",
          "language",
          "downloaded_bytes",
          "total_bytes",
          "progress_percent",
          "model_path",
          "storage_path",
          "error_code",
          "message",
          "retryable",
          "cancelable"
        ],
        "title": "SpeechModelStatusResponse",
        "description": "Stable public state for one provider-managed speech model."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}